[PR target/115123] Fix testsuite fallout from sinking heuristic change
Code sinking is just semantic preserving code motions, so it's a lot like
scheduling in that code motions can change the vector configuration needed at
various program points.  That in turn can also change the number of vsetvls as
we may or may not be able to merge them after the code motions.

The sinking heuristics were twiddled several months ago resulting in a handful
of scan-asm failures.  This patch adjusts the tests appropriately fixing
pr115123 (P3 regression).

	PR target/115123
gcc/testsuite
	* gcc.target/riscv/rvv/base/pr114352-3.c: Adjust expected output.
	* gcc.target/riscv/rvv/vsetvl/avl_multiple-7.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_multiple-8.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-66.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-82.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-83.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-86.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-88.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-90.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-91.c: Likewise.
	* gcc.target/riscv/rvv/vsetvl/avl_single-92.c: Likewise.