OpenMP: Support functions for nested "begin declare variant"
This patch adds functions for variant name mangling and context selector
merging that are shared by the C and C++ front ends.

The OpenMP specification says that name mangling is supposed to encode
the context selector for the variant, but also provides for no way to
reference these functions directly by name or from a different
compilation unit.  It also gives no guidance on how dynamic selectors
might be encoded across compilation units.

The GCC implementation of this feature instead treats variant
functions as if they have no linkage and uses a simple counter to
generate names.

gcc/ChangeLog
	* omp-general.cc (omp_mangle_variant_name): New.
	(omp_check_for_duplicate_variant): New.
	(omp_copy_trait_set): New.
	(omp_trait_selectors_equivalent): New.
	(omp_combine_trait_sets): New.
	(omp_merge_context_selectors): New.
	* omp-general.h (omp_mangle_variant_name): Declare.
	(omp_check_for_duplicate_variant): Declare.
	(omp_merge_context_selectors): Declare.