AArch64: Fix GCC 13 backport of big.Little CPU detection [PR118800]
On the GCC-13 branch the backport caused a failure due to the branch not having
generic-armv8-a and also it still treating the generic cpu special.  This made
it return NULL when trying to find the default CPU.

In GCC 13 we still had multiple structures with the same information and in this
case aarch64_cpu_data was missing the generic CPU which is in all_cores.

This corrects it by using "generc" instead and also adding it to
aarch64_cpu_data.

gcc/ChangeLog:

	PR target/118800
	* config/aarch64/driver-aarch64.cc (DEFAULT_CPU): Use generic instead of
	generic-armv8-a.
	(aarch64_cpu_data): Add generic.

gcc/testsuite/ChangeLog:

	PR target/118800
	* gcc.target/aarch64/cpunative/native_cpu_34.c: Update order.